backend/requirements.txt
django==4.1.5
djangorestframework==3.14.0
python-dotenv==0.21.0
django-environ==0.9.0
whitenoise==6.3.0
psycopg2-binary==2.9.5
